feat(modules): `deleteFile` support soft delete

+++ libs/modules/src/file.ts
@@ -1,5 +1,7 @@
 import { unlink } from "fs/promises";
+import trash from "trash";
 
-export async function deleteFile(path: string): Promise<void> {
+export async function deleteFile(path: string, softDelete = false): Promise<void> {
+  if (softDelete) return trash(path);
   return unlink(path);
 }
